+++ R/GLS.R
@@ -26,10 +26,26 @@
 #' @examples 
 #' data("dataDMA") 
 #' genes_query <- Reduce(c, dataDMA)
-#' dataGLS <- GLS(genes = genes_query) 
+#' dataGLS <- GLS(genes = genes_query,
+#'                query_string = "AND cancer AND driver AND '1980/01/01'[Date - Publication] : '2023/01/01'[Date - Publication]")
 GLS <- function(genes, 
                 query_string = "AND cancer AND driver",
 		max_records = 20) {
+
+  # Check user input
+  
+  if (!is.character(genes)) {
+    stop("Genes must be a character vector containing gene names to search
+         in PubMed")
+  }
+  
+  if (!is.character(query_string)) {
+    stop("The query string must be a character vector")
+  }
+  
+  if (!is.numeric(max_records)) {
+    stop("The maximum number of records to retrieve must be numeric")
+  }
   
   # Initialize empty tibble to store results
   pubmed_mining <- tibble()